+++ src/common/components/Navigation/SearchHeader/SearchHeader.tsx
@@ -3,8 +3,8 @@ import { CSSTransition } from "react-transition-group";
 import { css } from "twin.macro";
 
 import { useScrollDirection } from "@/common/hooks";
-import { SearchInput } from "@/components/search";
 import { TagCategory } from "@/components/tags";
+import { SearchInput } from "@/features/search/components";
 
 import { BackButton } from "../BackButton";
