--- src/components/search/SearchRecent/index.ts
@@ -1,1 +1,0 @@
-export * from "./SearchRecent";
