--- src/components/search/SearchResult/index.ts
@@ -1,1 +1,0 @@
-export * from "./SearchResultList";
