--- src/components/search/index.ts
@@ -1,5 +1,0 @@
-export * from "./SearchInput";
-export * from "./SearchItem";
-export * from "./SearchPopular";
-export * from "./SearchRecent";
-export * from "./SearchResult";
