+++ src/features/search/components/SearchInput/SearchInput.stories.tsx
