+++ src/features/search/components/SearchInput/index.ts
