+++ src/features/search/components/SearchResultList.tsx
@@ -4,7 +4,8 @@ import type { RecentSearch } from "@/application/hooks";
 import { useGetTagSearch } from "@/application/hooks";
 import { Icon } from "@/common/components/Icon";
 import { PATH } from "@/common/utils";
-import { SearchItem } from "@/components/search";
+
+import { SearchItem } from "./SearchItem";
 
 interface Prop {
   value: string;
