+++ src/features/search/components/Skeleton/index.ts
